+++ src/components/Inputs/Input.svelte
@@ -1,6 +1,8 @@
 <script lang="ts">
 	import type { HTMLInputTypeAttribute } from 'svelte/elements';
 
+	let className = '';
+
 	export let id: string;
 	export let label = '';
 	export let value: any = '';
@@ -8,7 +10,7 @@
 	export let type: HTMLInputTypeAttribute = 'text';
 	export let inputClass = '';
 	export let labelClass = '';
-	let className = '';
+	export let groupId = '';
 	export { className as class };
 
 	function typeAction(node: HTMLInputElement) {
@@ -46,7 +48,8 @@
 			type="radio"
 			{id}
 			{placeholder}
-			bind:group={value}
+			name={groupId.toString()}
+			bind:value
 			class={`mt-1 w-full outline-none py-3 px-3 rounded-[8px] border-gray-200 shadow-custom dark:shadow-custom-dark dark:border-gray-700 dark:bg-gray-800 dark:text-white sm:text-sm ${inputClass}`}
 		/>
 	{/if}